feat(meta): allow extra fields in item
these are now configured as vocabularies with specific field_type

SDESK-1679 SDESK-1636
